+++ lib/byebug/attacher.rb
@@ -8,8 +8,6 @@ module Byebug
   # Starts byebug, and stops at the first line of user's code.
   #
   def self.attach
-    require "byebug/core"
-
     unless started?
       self.mode = :attached
 
@@ -35,7 +33,9 @@ end
 #
 module Kernel
   def byebug
-    Byebug.attach
+    require "byebug/core"
+
+    Byebug.attach unless Byebug.mode == :off
   end
 
   def remote_byebug(host = "localhost", port = nil)